\name{extractoR.namespace-package}
\alias{extractoR.namespace-package}
\alias{extractoR.namespace}
\docType{package}
\encoding{UTF-8}
\title{CRAN packages metadata extractor (extract)}
\description{
  extractoR.namespace is part of extractoR which is a set of R packages
  that can be used to fetch, extract and query R packages metadata.
  extractoR.namespace parses NAMESPACE files from R packages.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}
